dom/svg/test/test_zoom.xhtml

<!DOCTYPE html>
<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=547596
-->  
<head>
  <title>Test for Bug 547596</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=547596">Mozilla Bug 547596</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>

  <iframe id="svg" src="zoom-helper.svg"></iframe>

  <script>
    SimpleTest.waitForExplicitFinish();

    function runTests() {
      var doc = $("svg").contentWindow.document;
      var root = doc.getElementById("root");
      root.currentScale *= 2;
    }

    function zoomHandler(evt) {
      ok(true, "zoom handler should be called on zoom");
      SimpleTest.finish();
    }

    window.addEventListener("load", runTests, false);
  </script>
</body>
</html>
